Remove unused hooks & refactor routes into module
